Our Ace Arsenal Extended Configs are getting unwieldy
