+++ addons/ace_arsenal_extended/data/aax_models/weapons/armour_custom/marines2_armour/config.hpp
@@ -1,0 +1,1 @@
+#include "group1/config.hpp"
