+++ addons/ace_arsenal_extended/data/aax_models/weapons/armour_custom/marines2_helmets/config.hpp
