+++ addons/ace_arsenal_extended/data/aax_models/weapons/armour_custom/odst2_armour/config.hpp
