+++ addons/ace_arsenal_extended/data/aax_models/weapons/armour_custom/odst2_helmets/config.hpp
@@ -1,0 +1,1 @@
+#include "group1\config.hpp"
