+++ addons/ace_arsenal_extended/data/aax_models/weapons/armour_custom/specialist_helmets/config.hpp
@@ -1,0 +1,1 @@
+#include "odst_group1\config.hpp"
